SearchKit doesn't let us search while we're flushing the index, so only fiush (in the background) when we close the search window, not immediately after we index.
There is a potential to appear to hang on quit here (when saving large changes to the searchIndex), so we need to devise some way to message the user this is happening if it's preventing quit.

The typical case is an incremental addition to the existing search index which takes next to no time at all.

--HG--
branch : AILoggerWithBlocks